Refactor: preserved route hierachy
