+++ src/components/common/index.jsx
@@ -1,9 +1,9 @@
-export { default as Layout } from "./layout";
 export { default as ScrollToTop } from "./scroll-to-top";
 
 export * from "./accordion";
 export * from "./buttons";
 export * from "./blocks";
+export * from "./layout";
 export * from "./tabs";
 export * from "./terminal";
 export * from "./typography";
